extract cell operations.
