+++ src/vs/workbench/contrib/notebook/browser/controller/coreActions.ts
@@ -4,29 +4,21 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { URI, UriComponents } from 'vs/base/common/uri';
-import { getIconClasses } from 'vs/editor/common/services/getIconClasses';
-import { IModelService } from 'vs/editor/common/services/modelService';
-import { IModeService } from 'vs/editor/common/services/modeService';
 import { localize } from 'vs/nls';
-import { Action2, IAction2Options, MenuId, MenuRegistry, registerAction2 } from 'vs/platform/actions/common/actions';
+import { Action2, IAction2Options, MenuId, MenuRegistry } from 'vs/platform/actions/common/actions';
 import { ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
-import { IQuickInputService, IQuickPickItem, QuickPickInput } from 'vs/platform/quickinput/common/quickInput';
-import { BaseCellRenderTemplate, CellEditState, getNotebookEditorFromEditorPane, IActiveNotebookEditor, ICellViewModel, NOTEBOOK_CELL_TYPE, NOTEBOOK_EDITOR_EDITABLE, NOTEBOOK_EDITOR_FOCUSED, NOTEBOOK_IS_ACTIVE_EDITOR, NOTEBOOK_KERNEL_COUNT, CHANGE_CELL_LANGUAGE, NOTEBOOK_CELL_EXECUTING, NOTEBOOK_MISSING_KERNEL_EXTENSION, cellRangeToViewCells } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
-import { CellEditType, CellKind, ICellEditOperation, SelectionStateType, ICellReplaceEdit } from 'vs/workbench/contrib/notebook/common/notebookCommon';
+import { BaseCellRenderTemplate, getNotebookEditorFromEditorPane, IActiveNotebookEditor, ICellViewModel, NOTEBOOK_EDITOR_EDITABLE, NOTEBOOK_EDITOR_FOCUSED, NOTEBOOK_IS_ACTIVE_EDITOR, NOTEBOOK_KERNEL_COUNT, cellRangeToViewCells } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
 import { ICellRange, isICellRange } from 'vs/workbench/contrib/notebook/common/notebookRange';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { IEditorCommandsContext } from 'vs/workbench/common/editor';
 import { INotebookEditorService } from 'vs/workbench/contrib/notebook/browser/notebookEditorService';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { WorkbenchActionExecutedClassification, WorkbenchActionExecutedEvent } from 'vs/base/common/actions';
-import { NotebookViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/notebookViewModel';
 import { flatten } from 'vs/base/common/arrays';
-import { Mimes } from 'vs/base/common/mime';
 import { TypeConstraint } from 'vs/base/common/types';
 import { IJSONSchema } from 'vs/base/common/jsonSchema';
-import { ILanguageDetectionService } from 'vs/workbench/services/languageDetection/common/languageDetectionWorkerService';
 import { MarshalledId } from 'vs/base/common/marshalling';
 
 // Kernel Command
@@ -286,18 +278,6 @@ export abstract class NotebookCellAction<T = INotebookCellActionContext> extends
 	abstract override runWithContext(accessor: ServicesAccessor, context: INotebookCellActionContext): Promise<void>;
 }
 
-// If this changes, update getCodeCellExecutionContextKeyService to match
-export const executeCondition = ContextKeyExpr.and(
-	NOTEBOOK_CELL_TYPE.isEqualTo('code'),
-	ContextKeyExpr.or(
-		ContextKeyExpr.greater(NOTEBOOK_KERNEL_COUNT.key, 0),
-		NOTEBOOK_MISSING_KERNEL_EXTENSION
-	));
-
-export const executeThisCellCondition = ContextKeyExpr.and(
-	executeCondition,
-	NOTEBOOK_CELL_EXECUTING.toNegated());
-
 export const executeNotebookCondition = ContextKeyExpr.greater(NOTEBOOK_KERNEL_COUNT.key, 0);
 
 interface IMultiCellArgs {
@@ -448,357 +428,4 @@ MenuRegistry.appendMenuItem(MenuId.EditorContext, {
 	title: localize('notebookMenu.cellTitle', "Notebook Cell"),
 	group: CellOverflowToolbarGroups.Insert,
 	when: NOTEBOOK_EDITOR_FOCUSED
-});
-
-export async function changeCellToKind(kind: CellKind, context: INotebookActionContext, language?: string, mime?: string): Promise<void> {
-	const { notebookEditor } = context;
-	if (!notebookEditor.viewModel) {
-		return;
-	}
-
-	if (notebookEditor.viewModel.options.isReadOnly) {
-		return;
-	}
-
-	if (context.ui && context.cell) {
-		// action from UI
-		const { cell } = context;
-
-		if (cell.cellKind === kind) {
-			return;
-		}
-
-		const text = cell.getText();
-		const idx = notebookEditor.viewModel.getCellIndex(cell);
-
-		if (language === undefined) {
-			const availableLanguages = notebookEditor.activeKernel?.supportedLanguages ?? [];
-			language = availableLanguages[0] ?? 'plaintext';
-		}
-
-		notebookEditor.textModel.applyEdits([
-			{
-				editType: CellEditType.Replace,
-				index: idx,
-				count: 1,
-				cells: [{
-					cellKind: kind,
-					source: text,
-					language: language!,
-					mime: mime ?? cell.mime,
-					outputs: cell.model.outputs,
-					metadata: cell.metadata,
-				}]
-			}
-		], true, {
-			kind: SelectionStateType.Index,
-			focus: notebookEditor.getFocus(),
-			selections: notebookEditor.getSelections()
-		}, () => {
-			return {
-				kind: SelectionStateType.Index,
-				focus: notebookEditor.getFocus(),
-				selections: notebookEditor.getSelections()
-			};
-		}, undefined, true);
-		const newCell = notebookEditor.viewModel.cellAt(idx);
-
-		if (!newCell) {
-			return;
-		}
-
-		notebookEditor.focusNotebookCell(newCell, cell.getEditState() === CellEditState.Editing ? 'editor' : 'container');
-	} else if (context.selectedCells) {
-		const selectedCells = context.selectedCells;
-		const rawEdits: ICellEditOperation[] = [];
-
-		selectedCells.forEach(cell => {
-			if (cell.cellKind === kind) {
-				return;
-			}
-			const text = cell.getText();
-			const idx = notebookEditor.viewModel.getCellIndex(cell);
-
-			if (language === undefined) {
-				const availableLanguages = notebookEditor.activeKernel?.supportedLanguages ?? [];
-				language = availableLanguages[0] ?? 'plaintext';
-			}
-
-			rawEdits.push(
-				{
-					editType: CellEditType.Replace,
-					index: idx,
-					count: 1,
-					cells: [{
-						cellKind: kind,
-						source: text,
-						language: language!,
-						mime: mime ?? cell.mime,
-						outputs: cell.model.outputs,
-						metadata: cell.metadata,
-					}]
-				}
-			);
-		});
-
-		notebookEditor.textModel.applyEdits(rawEdits, true, {
-			kind: SelectionStateType.Index,
-			focus: notebookEditor.getFocus(),
-			selections: notebookEditor.getSelections()
-		}, () => {
-			return {
-				kind: SelectionStateType.Index,
-				focus: notebookEditor.getFocus(),
-				selections: notebookEditor.getSelections()
-			};
-		}, undefined, true);
-	}
-}
-
-export function runDeleteAction(viewModel: NotebookViewModel, cell: ICellViewModel) {
-	const selections = viewModel.getSelections();
-	const targetCellIndex = viewModel.getCellIndex(cell);
-	const containingSelection = selections.find(selection => selection.start <= targetCellIndex && targetCellIndex < selection.end);
-
-	if (containingSelection) {
-		const edits: ICellReplaceEdit[] = selections.reverse().map(selection => ({
-			editType: CellEditType.Replace, index: selection.start, count: selection.end - selection.start, cells: []
-		}));
-
-		const nextCellAfterContainingSelection = viewModel.cellAt(containingSelection.end);
-
-		viewModel.notebookDocument.applyEdits(edits, true, { kind: SelectionStateType.Index, focus: viewModel.getFocus(), selections: viewModel.getSelections() }, () => {
-			if (nextCellAfterContainingSelection) {
-				const cellIndex = viewModel.notebookDocument.cells.findIndex(cell => cell.handle === nextCellAfterContainingSelection.handle);
-				return { kind: SelectionStateType.Index, focus: { start: cellIndex, end: cellIndex + 1 }, selections: [{ start: cellIndex, end: cellIndex + 1 }] };
-			} else {
-				if (viewModel.notebookDocument.length) {
-					const lastCellIndex = viewModel.notebookDocument.length - 1;
-					return { kind: SelectionStateType.Index, focus: { start: lastCellIndex, end: lastCellIndex + 1 }, selections: [{ start: lastCellIndex, end: lastCellIndex + 1 }] };
-
-				} else {
-					return { kind: SelectionStateType.Index, focus: { start: 0, end: 0 }, selections: [{ start: 0, end: 0 }] };
-				}
-			}
-		}, undefined);
-	} else {
-		const focus = viewModel.getFocus();
-		const edits: ICellReplaceEdit[] = [{
-			editType: CellEditType.Replace, index: targetCellIndex, count: 1, cells: []
-		}];
-
-		let finalSelections: ICellRange[] = [];
-		for (let i = 0; i < selections.length; i++) {
-			const selection = selections[i];
-
-			if (selection.end <= targetCellIndex) {
-				finalSelections.push(selection);
-			} else if (selection.start > targetCellIndex) {
-				finalSelections.push({ start: selection.start - 1, end: selection.end - 1 });
-			} else {
-				finalSelections.push({ start: targetCellIndex, end: targetCellIndex + 1 });
-			}
-		}
-
-		if (viewModel.cellAt(focus.start) === cell) {
-			// focus is the target, focus is also not part of any selection
-			const newFocus = focus.end === viewModel.length ? { start: focus.start - 1, end: focus.end - 1 } : focus;
-
-			viewModel.notebookDocument.applyEdits(edits, true, { kind: SelectionStateType.Index, focus: viewModel.getFocus(), selections: viewModel.getSelections() }, () => ({
-				kind: SelectionStateType.Index, focus: newFocus, selections: finalSelections
-			}), undefined);
-		} else {
-			// users decide to delete a cell out of current focus/selection
-			const newFocus = focus.start > targetCellIndex ? { start: focus.start - 1, end: focus.end - 1 } : focus;
-
-			viewModel.notebookDocument.applyEdits(edits, true, { kind: SelectionStateType.Index, focus: viewModel.getFocus(), selections: viewModel.getSelections() }, () => ({
-				kind: SelectionStateType.Index, focus: newFocus, selections: finalSelections
-			}), undefined);
-		}
-	}
-}
-
-interface ILanguagePickInput extends IQuickPickItem {
-	languageId: string;
-	description: string;
-}
-
-interface IChangeCellContext extends INotebookCellActionContext {
-	// TODO@rebornix : `cells`
-	// range: ICellRange;
-	language?: string;
-}
-
-registerAction2(class ChangeCellLanguageAction extends NotebookCellAction<ICellRange> {
-	constructor() {
-		super({
-			id: CHANGE_CELL_LANGUAGE,
-			title: localize('changeLanguage', 'Change Cell Language'),
-			description: {
-				description: localize('changeLanguage', 'Change Cell Language'),
-				args: [
-					{
-						name: 'range',
-						description: 'The cell range',
-						schema: {
-							'type': 'object',
-							'required': ['start', 'end'],
-							'properties': {
-								'start': {
-									'type': 'number'
-								},
-								'end': {
-									'type': 'number'
-								}
-							}
-						}
-					},
-					{
-						name: 'language',
-						description: 'The target cell language',
-						schema: {
-							'type': 'string'
-						}
-					}
-				]
-			}
-		});
-	}
-
-	protected override getCellContextFromArgs(accessor: ServicesAccessor, context?: ICellRange, ...additionalArgs: any[]): IChangeCellContext | undefined {
-		if (!context || typeof context.start !== 'number' || typeof context.end !== 'number' || context.start >= context.end) {
-			return;
-		}
-
-		const language = additionalArgs.length && typeof additionalArgs[0] === 'string' ? additionalArgs[0] : undefined;
-		const activeEditorContext = this.getEditorContextFromArgsOrActive(accessor);
-
-		if (!activeEditorContext || !activeEditorContext.notebookEditor.viewModel || context.start >= activeEditorContext.notebookEditor.viewModel.length) {
-			return;
-		}
-
-		// TODO@rebornix, support multiple cells
-		return {
-			notebookEditor: activeEditorContext.notebookEditor,
-			cell: activeEditorContext.notebookEditor.viewModel.cellAt(context.start)!,
-			language
-		};
-	}
-
-
-	async runWithContext(accessor: ServicesAccessor, context: IChangeCellContext): Promise<void> {
-		if (context.language) {
-			await this.setLanguage(context, context.language);
-		} else {
-			await this.showLanguagePicker(accessor, context);
-		}
-	}
-
-	private async showLanguagePicker(accessor: ServicesAccessor, context: IChangeCellContext) {
-		const topItems: ILanguagePickInput[] = [];
-		const mainItems: ILanguagePickInput[] = [];
-
-		const modeService = accessor.get(IModeService);
-		const modelService = accessor.get(IModelService);
-		const quickInputService = accessor.get(IQuickInputService);
-		const languageDetectionService = accessor.get(ILanguageDetectionService);
-
-		const providerLanguages = new Set([
-			...(context.notebookEditor.activeKernel?.supportedLanguages ?? modeService.getRegisteredModes()),
-			'markdown'
-		]);
-
-		providerLanguages.forEach(languageId => {
-			let description: string;
-			if (context.cell.cellKind === CellKind.Markup ? (languageId === 'markdown') : (languageId === context.cell.language)) {
-				description = localize('languageDescription', "({0}) - Current Language", languageId);
-			} else {
-				description = localize('languageDescriptionConfigured', "({0})", languageId);
-			}
-
-			const languageName = modeService.getLanguageName(languageId);
-			if (!languageName) {
-				// Notebook has unrecognized language
-				return;
-			}
-
-			const item = <ILanguagePickInput>{
-				label: languageName,
-				iconClasses: getIconClasses(modelService, modeService, this.getFakeResource(languageName, modeService)),
-				description,
-				languageId
-			};
-
-			if (languageId === 'markdown' || languageId === context.cell.language) {
-				topItems.push(item);
-			} else {
-				mainItems.push(item);
-			}
-		});
-
-		mainItems.sort((a, b) => {
-			return a.description.localeCompare(b.description);
-		});
-
-		// Offer to "Auto Detect"
-		const autoDetectMode: IQuickPickItem = {
-			label: localize('autoDetect', "Auto Detect")
-		};
-
-		const picks: QuickPickInput[] = [
-			autoDetectMode,
-			{ type: 'separator', label: localize('languagesPicks', "languages (identifier)") },
-			...topItems,
-			{ type: 'separator' },
-			...mainItems
-		];
-
-		const selection = await quickInputService.pick(picks, { placeHolder: localize('pickLanguageToConfigure', "Select Language Mode") }) as ILanguagePickInput | undefined;
-		let languageId = selection === autoDetectMode
-			? await languageDetectionService.detectLanguage(context.cell.uri)
-			: selection?.languageId;
-
-		if (languageId) {
-			await this.setLanguage(context, languageId);
-		}
-	}
-
-	private async setLanguage(context: IChangeCellContext, languageId: string) {
-		if (languageId === 'markdown' && context.cell?.language !== 'markdown') {
-			const idx = context.notebookEditor.viewModel.getCellIndex(context.cell);
-			await changeCellToKind(CellKind.Markup, { cell: context.cell, notebookEditor: context.notebookEditor }, 'markdown', Mimes.markdown);
-			const newCell = context.notebookEditor.viewModel.cellAt(idx);
-
-			if (newCell) {
-				context.notebookEditor.focusNotebookCell(newCell, 'editor');
-			}
-		} else if (languageId !== 'markdown' && context.cell?.cellKind === CellKind.Markup) {
-			await changeCellToKind(CellKind.Code, { cell: context.cell, notebookEditor: context.notebookEditor }, languageId);
-		} else {
-			const index = context.notebookEditor.textModel.cells.indexOf(context.cell.model);
-			context.notebookEditor.textModel.applyEdits(
-				[{ editType: CellEditType.CellLanguage, index, language: languageId }],
-				true, undefined, () => undefined, undefined
-			);
-		}
-	}
-
-	/**
-	 * Copied from editorStatus.ts
-	 */
-	private getFakeResource(lang: string, modeService: IModeService): URI | undefined {
-		let fakeResource: URI | undefined;
-
-		const extensions = modeService.getExtensions(lang);
-		if (extensions?.length) {
-			fakeResource = URI.file(extensions[0]);
-		} else {
-			const filenames = modeService.getFilenames(lang);
-			if (filenames?.length) {
-				fakeResource = URI.file(filenames[0]);
-			}
-		}
-
-		return fakeResource;
-	}
 });
