+++ src/vs/workbench/contrib/notebook/browser/controller/editActions.ts
@@ -4,7 +4,12 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
+import { Mimes } from 'vs/base/common/mime';
+import { URI } from 'vs/base/common/uri';
 import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
+import { getIconClasses } from 'vs/editor/common/services/getIconClasses';
+import { IModelService } from 'vs/editor/common/services/modelService';
+import { IModeService } from 'vs/editor/common/services/modeService';
 import { localize } from 'vs/nls';
 import { MenuId, MenuItemAction, registerAction2 } from 'vs/platform/actions/common/actions';
 import { ICommandService } from 'vs/platform/commands/common/commands';
@@ -12,10 +17,14 @@ import { ContextKeyExpr, IContextKeyService } from 'vs/platform/contextkey/commo
 import { InputFocusedContext, InputFocusedContextKey } from 'vs/platform/contextkey/common/contextkeys';
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
-import { CellToolbarOrder, CELL_TITLE_CELL_GROUP_ID, CELL_TITLE_OUTPUT_GROUP_ID, executeNotebookCondition, INotebookActionContext, INotebookCellActionContext, NotebookAction, NotebookCellAction, NOTEBOOK_EDITOR_WIDGET_ACTION_WEIGHT, runDeleteAction } from 'vs/workbench/contrib/notebook/browser/controller/coreActions';
-import { CellEditState, NOTEBOOK_CELL_EDITABLE, NOTEBOOK_CELL_HAS_OUTPUTS, NOTEBOOK_CELL_LIST_FOCUSED, NOTEBOOK_CELL_MARKDOWN_EDIT_MODE, NOTEBOOK_CELL_TYPE, NOTEBOOK_EDITOR_EDITABLE, NOTEBOOK_EDITOR_FOCUSED, NOTEBOOK_HAS_OUTPUTS, NOTEBOOK_IS_ACTIVE_EDITOR, NOTEBOOK_USE_CONSOLIDATED_OUTPUT_BUTTON, QUIT_EDIT_CELL_COMMAND_ID } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
+import { IQuickInputService, IQuickPickItem, QuickPickInput } from 'vs/platform/quickinput/common/quickInput';
+import { changeCellToKind, runDeleteAction } from 'vs/workbench/contrib/notebook/browser/controller/cellOperations';
+import { CellToolbarOrder, CELL_TITLE_CELL_GROUP_ID, CELL_TITLE_OUTPUT_GROUP_ID, executeNotebookCondition, INotebookActionContext, INotebookCellActionContext, NotebookAction, NotebookCellAction, NOTEBOOK_EDITOR_WIDGET_ACTION_WEIGHT } from 'vs/workbench/contrib/notebook/browser/controller/coreActions';
+import { CellEditState, CHANGE_CELL_LANGUAGE, NOTEBOOK_CELL_EDITABLE, NOTEBOOK_CELL_HAS_OUTPUTS, NOTEBOOK_CELL_LIST_FOCUSED, NOTEBOOK_CELL_MARKDOWN_EDIT_MODE, NOTEBOOK_CELL_TYPE, NOTEBOOK_EDITOR_EDITABLE, NOTEBOOK_EDITOR_FOCUSED, NOTEBOOK_HAS_OUTPUTS, NOTEBOOK_IS_ACTIVE_EDITOR, NOTEBOOK_USE_CONSOLIDATED_OUTPUT_BUTTON, QUIT_EDIT_CELL_COMMAND_ID } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
 import * as icons from 'vs/workbench/contrib/notebook/browser/notebookIcons';
 import { CellEditType, CellKind, ICellEditOperation, NotebookCellExecutionState } from 'vs/workbench/contrib/notebook/common/notebookCommon';
+import { ICellRange } from 'vs/workbench/contrib/notebook/common/notebookRange';
+import { ILanguageDetectionService } from 'vs/workbench/services/languageDetection/common/languageDetectionWorkerService';
 
 const CLEAR_ALL_CELLS_OUTPUTS_COMMAND_ID = 'notebook.clearAllCellsOutputs';
 const EDIT_CELL_COMMAND_ID = 'notebook.cell.edit';
@@ -280,5 +289,191 @@ registerAction2(class ClearAllCellOutputsAction extends NotebookAction {
 		if (clearExecutionMetadataEdits.length) {
 			context.notebookEditor.textModel.applyEdits(clearExecutionMetadataEdits, true, undefined, () => undefined, undefined);
 		}
+	}
+});
+
+
+interface ILanguagePickInput extends IQuickPickItem {
+	languageId: string;
+	description: string;
+}
+
+interface IChangeCellContext extends INotebookCellActionContext {
+	// TODO@rebornix : `cells`
+	// range: ICellRange;
+	language?: string;
+}
+
+registerAction2(class ChangeCellLanguageAction extends NotebookCellAction<ICellRange> {
+	constructor() {
+		super({
+			id: CHANGE_CELL_LANGUAGE,
+			title: localize('changeLanguage', 'Change Cell Language'),
+			description: {
+				description: localize('changeLanguage', 'Change Cell Language'),
+				args: [
+					{
+						name: 'range',
+						description: 'The cell range',
+						schema: {
+							'type': 'object',
+							'required': ['start', 'end'],
+							'properties': {
+								'start': {
+									'type': 'number'
+								},
+								'end': {
+									'type': 'number'
+								}
+							}
+						}
+					},
+					{
+						name: 'language',
+						description: 'The target cell language',
+						schema: {
+							'type': 'string'
+						}
+					}
+				]
+			}
+		});
+	}
+
+	protected override getCellContextFromArgs(accessor: ServicesAccessor, context?: ICellRange, ...additionalArgs: any[]): IChangeCellContext | undefined {
+		if (!context || typeof context.start !== 'number' || typeof context.end !== 'number' || context.start >= context.end) {
+			return;
+		}
+
+		const language = additionalArgs.length && typeof additionalArgs[0] === 'string' ? additionalArgs[0] : undefined;
+		const activeEditorContext = this.getEditorContextFromArgsOrActive(accessor);
+
+		if (!activeEditorContext || !activeEditorContext.notebookEditor.viewModel || context.start >= activeEditorContext.notebookEditor.viewModel.length) {
+			return;
+		}
+
+		// TODO@rebornix, support multiple cells
+		return {
+			notebookEditor: activeEditorContext.notebookEditor,
+			cell: activeEditorContext.notebookEditor.viewModel.cellAt(context.start)!,
+			language
+		};
+	}
+
+
+	async runWithContext(accessor: ServicesAccessor, context: IChangeCellContext): Promise<void> {
+		if (context.language) {
+			await this.setLanguage(context, context.language);
+		} else {
+			await this.showLanguagePicker(accessor, context);
+		}
+	}
+
+	private async showLanguagePicker(accessor: ServicesAccessor, context: IChangeCellContext) {
+		const topItems: ILanguagePickInput[] = [];
+		const mainItems: ILanguagePickInput[] = [];
+
+		const modeService = accessor.get(IModeService);
+		const modelService = accessor.get(IModelService);
+		const quickInputService = accessor.get(IQuickInputService);
+		const languageDetectionService = accessor.get(ILanguageDetectionService);
+
+		const providerLanguages = new Set([
+			...(context.notebookEditor.activeKernel?.supportedLanguages ?? modeService.getRegisteredModes()),
+			'markdown'
+		]);
+
+		providerLanguages.forEach(languageId => {
+			let description: string;
+			if (context.cell.cellKind === CellKind.Markup ? (languageId === 'markdown') : (languageId === context.cell.language)) {
+				description = localize('languageDescription', "({0}) - Current Language", languageId);
+			} else {
+				description = localize('languageDescriptionConfigured', "({0})", languageId);
+			}
+
+			const languageName = modeService.getLanguageName(languageId);
+			if (!languageName) {
+				// Notebook has unrecognized language
+				return;
+			}
+
+			const item = <ILanguagePickInput>{
+				label: languageName,
+				iconClasses: getIconClasses(modelService, modeService, this.getFakeResource(languageName, modeService)),
+				description,
+				languageId
+			};
+
+			if (languageId === 'markdown' || languageId === context.cell.language) {
+				topItems.push(item);
+			} else {
+				mainItems.push(item);
+			}
+		});
+
+		mainItems.sort((a, b) => {
+			return a.description.localeCompare(b.description);
+		});
+
+		// Offer to "Auto Detect"
+		const autoDetectMode: IQuickPickItem = {
+			label: localize('autoDetect', "Auto Detect")
+		};
+
+		const picks: QuickPickInput[] = [
+			autoDetectMode,
+			{ type: 'separator', label: localize('languagesPicks', "languages (identifier)") },
+			...topItems,
+			{ type: 'separator' },
+			...mainItems
+		];
+
+		const selection = await quickInputService.pick(picks, { placeHolder: localize('pickLanguageToConfigure', "Select Language Mode") }) as ILanguagePickInput | undefined;
+		let languageId = selection === autoDetectMode
+			? await languageDetectionService.detectLanguage(context.cell.uri)
+			: selection?.languageId;
+
+		if (languageId) {
+			await this.setLanguage(context, languageId);
+		}
+	}
+
+	private async setLanguage(context: IChangeCellContext, languageId: string) {
+		if (languageId === 'markdown' && context.cell?.language !== 'markdown') {
+			const idx = context.notebookEditor.viewModel.getCellIndex(context.cell);
+			await changeCellToKind(CellKind.Markup, { cell: context.cell, notebookEditor: context.notebookEditor }, 'markdown', Mimes.markdown);
+			const newCell = context.notebookEditor.viewModel.cellAt(idx);
+
+			if (newCell) {
+				context.notebookEditor.focusNotebookCell(newCell, 'editor');
+			}
+		} else if (languageId !== 'markdown' && context.cell?.cellKind === CellKind.Markup) {
+			await changeCellToKind(CellKind.Code, { cell: context.cell, notebookEditor: context.notebookEditor }, languageId);
+		} else {
+			const index = context.notebookEditor.textModel.cells.indexOf(context.cell.model);
+			context.notebookEditor.textModel.applyEdits(
+				[{ editType: CellEditType.CellLanguage, index, language: languageId }],
+				true, undefined, () => undefined, undefined
+			);
+		}
+	}
+
+	/**
+	 * Copied from editorStatus.ts
+	 */
+	private getFakeResource(lang: string, modeService: IModeService): URI | undefined {
+		let fakeResource: URI | undefined;
+
+		const extensions = modeService.getExtensions(lang);
+		if (extensions?.length) {
+			fakeResource = URI.file(extensions[0]);
+		} else {
+			const filenames = modeService.getFilenames(lang);
+			if (filenames?.length) {
+				fakeResource = URI.file(filenames[0]);
+			}
+		}
+
+		return fakeResource;
 	}
 });
